apps/hrdata/serializers.py
from rest_framework import serializers
from .models import Topic, Tag, Document, Region

class TopicSerializer(serializers.ModelSerializer):
    class Meta:
        model = Topic
        fields = "__all__"

class TagSerializer(serializers.ModelSerializer):
    class Meta:
        model = Tag
        fields = "__all__"

class RegionSerializer(serializers.ModelSerializer):
    class Meta:
        model = Region
        fields = "__all__"

class DocumentSerializer(serializers.ModelSerializer):
    class Meta:
        model = Document
        fields = "__all__"
        # exclude = ['last_updated_by']
        read_only_fields = ('last_updated_by',)
    
